install.sh : doesn't overwrite Conf file anymore

+++ src/install.sh
@@ -4,7 +4,10 @@ if [ ! -d /etc/eztv-downloader/shows ]; then
 	mkdir -p /etc/eztv-downloader/shows
 fi
 
-cp Conf /etc/eztv-downloader
+if [ ! -e /etc/eztv-downloader/Conf ]; then
+    cp Conf /etc/eztv-downloader
+fi
+
 cp eztv-checkEpisodes.py /usr/bin/
 cp eztv-addTVshow.py /usr/bin
 cp eztvLogger.py /usr/bin
